ClockInfoDao

+++ app/src/main/java/com/example/pill_checker/data/ClockInfo.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "ClockInfo")
 data class ClockInfo(
-    @PrimaryKey val cid: Long,
+    @PrimaryKey val cid: Int,
     var morningHour: Int,
     var morningMin: Int,
     var lunchHour: Int,
